backend/requirements.txt
fastapi
uvicorn
supabase
pydantic
PyPDF2
python-dotenv
langchain
langchain-openai
playwright # after installation execute in cmd: playwright install chromium